fix: add period-stats command
